통계 진행중...

+++ SecondView/SecondVCExtension/TableView + Delegate.swift
@@ -23,6 +23,7 @@ extension SecondViewController: UITableViewDelegate, UITableViewDataSource {
             return UITableViewCell()
         }
         
+        cell.selectionStyle = .none
         cell.scheduleTitle.text = dayTasks[indexPath.row].schedule
         cell.scheduleTime.text = "\(dayTasks[indexPath.row].startTime)~\(dayTasks[indexPath.row].endTime)"
         dayTasks[indexPath.row].scheduleSuccess == true ? cell.checkButton.setImage(UIImage(systemName: "checkmark.square") , for: .normal) : cell.checkButton.setImage(UIImage(systemName: "x.square"), for: .normal)
@@ -30,24 +31,25 @@ extension SecondViewController: UITableViewDelegate, UITableViewDataSource {
         return cell
     }
     
+    // 스와이프 삭제
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        
-        if editingStyle == .delete {
-            repository.delete(item: dayTasks?[indexPath.row])
-        }
         
         let components = calendar.dateComponents([.year, .month], from: date)
         let startOfMonth = calendar.date(from: components)!
         
-        for i in 0..<repository.successSchedule(currentDate: startOfMonth).count {
-            scheduleCountDic.updateValue(repository.successScheduleNumber(key: repository.successSchedule(currentDate: startOfMonth)[i].schedule).count, forKey: repository.successSchedule(currentDate: startOfMonth)[i].schedule)
+        if dayTasks[indexPath.row].scheduleDate < now {
+            showAlertOnlyOk(title: "지나간 일정은 삭제할 수 없습니다.")
+        } else {
+            if editingStyle == .delete {
+                repository.delete(item: dayTasks?[indexPath.row])
+            }
+            
+            mainView.calendar.reloadData()
+            self.fetchRealm()
         }
-        
-        mainView.collectionView.reloadData()
-        mainView.calendar.reloadData()
-        self.fetchRealm()
     }
     
+    // 셀 선택
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
             
         let vc = ChangeScheduleViewController()
@@ -58,14 +60,17 @@ extension SecondViewController: UITableViewDelegate, UITableViewDataSource {
         vc.objectID = dayTasks[indexPath.row].objectID
         vc.receivedDate = dayTasks[indexPath.row].scheduleDate
         
-        // 2. 클로저 함수 정의
-        vc.okButtonActionHandler = {
-            self.mainView.tableView.reloadData()
+        if dayTasks[indexPath.row].scheduleDate < now {
+            showAlertOnlyOk(title: "지나간 일정은 수정할 수 없습니다.")
+        } else {
+            
+            // 2. 클로저 함수 정의
+            vc.okButtonActionHandler = {
+                self.mainView.tableView.reloadData()
+            }
+            
+            let nav = UINavigationController(rootViewController: vc)
+            present(nav, animated: true)
         }
-        
-        self.mainView.collectionView.reloadData()
-        
-        let nav = UINavigationController(rootViewController: vc)
-        present(nav, animated: true)
     }
 }
